deduplicate seaweedfsLowerBandIPIndex
